move build directory for trunk to work better. some cleanup
