+++ .trunk/configs/.isort.cfg
@@ -1,0 +1,2 @@
+[settings]
+profile=black
