+++ .trunk/trunk.yaml
@@ -2,11 +2,11 @@
 # To learn more about the format of this file, see https://docs.trunk.io/reference/trunk-yaml
 version: 0.1
 cli:
-  version: 1.17.1
+  version: 1.22.1
 plugins:
   sources:
     - id: trunk
-      ref: v1.2.6
+      ref: v1.5.0
       uri: https://github.com/trunk-io/plugins
 runtimes:
   enabled:
@@ -15,23 +15,29 @@ runtimes:
     - python@3.10.8
 lint:
   enabled:
-    - shellcheck@0.9.0
+    - taplo@0.8.1
+    - bandit@1.7.8
+    - black@24.4.2
+    - isort@5.13.2
+    - ruff@0.4.4
+    - shellcheck@0.10.0
     - shfmt@3.6.0
     - pragma-once
     - include-what-you-use@0.20
     - clang-tidy@16.0.3
     - clang-format@16.0.3
-    - trivy@0.46.1
-    - yamllint@1.32.0
-    - checkov@3.0.15
+    - trivy@0.51.1
+    - yamllint@1.35.1
+    - checkov@3.2.94
     - git-diff-check
-    - markdownlint@0.37.0
-    - prettier@3.0.3
-    - trufflehog@3.62.1
+    - markdownlint@0.40.0
+    - prettier@3.2.5
+    - trufflehog@3.76.3
   ignore:
     - linters: [ALL]
       paths:
         - .vscode/settings.json
+        - build/**
 actions:
   disabled:
     - trunk-announce
