+++ include/FieldUtils.h
@@ -1,0 +1,37 @@
+#pragma once
+
+#include <memory>
+#include <stk_mesh/base/BulkData.hpp>
+#include <stk_mesh/base/Field.hpp>
+#include <stk_mesh/base/GetNgpField.hpp>
+#include <stk_mesh/base/GetNgpMesh.hpp>
+#include <stk_mesh/base/MetaData.hpp>
+#include <stk_mesh/base/NgpField.hpp>
+#include <stk_mesh/base/NgpForEachEntity.hpp>
+#include <stk_mesh/base/NgpMesh.hpp>
+
+#include "FieldData.h"
+#include "MeshData.h"
+
+namespace aperi {
+
+inline stk::mesh::Field<double> *StkGetField(const FieldQueryData &field_query_data, stk::mesh::MetaData *meta_data) {
+    stk::topology::rank_t rank = field_query_data.rank == FieldDataRank::NODE ? stk::topology::NODE_RANK : stk::topology::ELEMENT_RANK;
+    stk::mesh::Field<double> *field = meta_data->get_field<double>(rank, field_query_data.name);
+    if (field == nullptr) {
+        throw std::runtime_error("Field " + field_query_data.name + " not found.");
+    }
+    stk::mesh::FieldState state = stk::mesh::StateNone;
+    if (field_query_data.state == FieldQueryState::N) {
+        state = stk::mesh::StateN;
+    } else if (field_query_data.state == FieldQueryState::NP1) {
+        state = stk::mesh::StateNP1;
+    } else {
+        if (field_query_data.state != FieldQueryState::None) {
+            throw std::runtime_error("Invalid field state");
+        }
+    }
+    return &field->field_of_state(state);
+}
+
+}  // namespace aperi
